Sprint8Features with Jpmml
This branch has the Jpmml support built on top of the Sprint8Features branch as of Dec 4, 2015.
This branch builds but has not been tested.

Until the final solution (the factory of factories is implemented) for factory objects, the ModelBase
and ModelBaseObj have been modified to include the model name and model version in several places.  This
has rippled back through the KamanjaMetadata and LearningEngine for now.

This branch will allow me to test the JPMML on top of the Sprint8Features changes while more integration
work is accomplished by Pokuri with his ModelInterfaceChanges branch... also on top of the Sprint8Features.