Fatafat218 Jpmml integration with new dev changes
The JpmmlAdapter and its factory have been brought into compliance with the
new interfaces found in dev.

What remains is to build a factory of factory for jpmml model factories.  One of these will be built
as a sibling project to trunk/FactoriesOfModelInstanceFactory/JarFactoryOfModelInstanceFactor for
jpmml next.  The integration point is the KamanjaMetadata where models are currently assumed to only
be "JAR" based.  A table of factory of factory instances will be built that will have as key the
model representation value ... any {JAR, JPMML} and a value of the corresponding factory of factory
instance.

Once implemented, a regression of the metadata api for jpmml and all other model types.

Then we test the factory of factories for jpmml and jar based representations followed by actual
engine processing tests .