Fatafat218 jpmml integration w/new model interfaces
1. implement 'def loadTypesFromAFile(input : String, userid : Option[String] = None): String'  in TypeService.scala
needed for pmml testing.
2. Gathered Pmml and Native model test cases into RegressionTestCases.md.
3. Moved several of the pmml models in the SampleApplication/Medical/Models/ to the attic.  These need some changes
to make them work properly... there is still one pmml model in the SampleApplications that get installed.
4. Modified pmml unit tests that test macros and missing values to use the NamespaceSearchPath instead of the older
UdfSearchPath.

Ingestion of pmml and native models work.  See the RegressionTestCases.md for commands needed to compile these models.

Tomorrow, the jpmml factory of factory loading and model execution is undertaken.  There is going to be a merge at the
end of the day as well to pick up the latest dev enhancements.