KAMANJA-343 ClusterInstall (top level) driver
Initial checkin.
Does not build.
Most functions with implementations.
No clean up on the ClusterInstall script to supply the parent dir and the new/old installation dir names YET

I thought it would be useful to allow you to see this Pokuri to see if this is what is required. I need to leave now.
I will be back online around 10pm.  Perhaps we can chat then.  This is still pretty rough and it has not been put through
the compiler yet... so cut me some slack.

I will make it build and provide implementation for the rest of it... hopefully before bedtime.