Changes for KAMANJA-331
Changes for https://ligadata.atlassian.net/browse/KAMANJA-331

1. Include multiple watch directories (separated by file.separator
character)
2. Include the Full File Name while Processing (earlier it was just the
file name, now because of multiple directories,the absolute path also
needs to be considered with the file name).
3. Change the way file names are stored in ZooKeeper (file path info
needs to be encoded/decoded)
4. Generate a random string (10 digits) for each file. Apache Commons
Lang 3 package used for this purpose. Checked the randomness for 30MM
strings and found no duplicates.
5. Inject File ID and Offset into the message (added in the beginning
based on a config flag)
6. Sniff the file for the file type (current scope limited to GZIP and
TXT files). Apache TIKA core is used for this purpose
7. Based on a config flag, emit the full FileName and the offset,
whenever there is a parsing issue