KAMANJA-343 node config extract from json
1) The information required by the cluster installer script is now being
extracted from the config file JSON.  All the top level JSON nodes have maps built for them.
2) The clusterInstallerDriver was in the ClusterInstall folder which was copied in total to the
installation directory.  It has been now moved up one and is a sibling with the ClusterInstall

To do:
1) write node config info to files and modify the call to cluster installer to use them instead of
calling the node info extract in the subscript.
2) detail the method that will validate the cluster environmnent throught the use of calls to Yousef's tools
3) make sure the substitution for the migrate config works
4) include the migrate jars directly in the clusterInstallerDriver project and use the new side-door methods
that Pokuri has provided in the migrate manager.
5) Pull the node info configuration extract class into a separate file. It may prove useful for other projects.