KAMANJA-343 InstallDriver Migrate,KamanjaClusterInstall
Integration with the KamanjaClusterInstall (three files formerly generated in KamanjaClusterInstall's call to NodeInfoExtract)
now supplied by the InstallerDriver.  NodeInfoExtract NO LONGER USED.
Integration with the MigrateManager. The InstallDriver log class implements the call back that is used by Migrate.
The MigrateManager directly used by the InstallDriver.
The easyInstallKamanja.sh script gets a new project for both the 2.10 and 2.11 sides (clusterInstallerDriver). Fat jars
produced for these.
The MigrateConfigTemplate.json file added to project for reference.

What is left:

1. I have not built the control loops to call Yousef's tool via the new scp/ssh/scp triad of scripts that are being built
to access the cluster nodes and discern the various information that we want to verify.
2. I will add that later this morning when I awake.  Pokuri has defined the info to be submitted and what will be returned.
3. More testing.