KAMANJA-343 getFileName fix
Make the getFileName more flexible.  Add boolean argument that checks the parent dir for existence
but allows the file name to not exist and add a boolean that insists the file (and indirectly) the
parent directory exists.

When the getComponent script fails, as a temp measure, trap the exception that is thrown, give
the physical dir name as the one computed from the ROOT_DIR property so testing can continue a bit
to the next problem.  This will be taken out once the script is being invoked and returning real
results.

Status:

The get component script is returning a 1 (non 0 is fail). The error is trapped (as described) so testing can continue.
The low level install script is invoked and returns a 127 error (non 0 is fail).

Question: Is the script that has been checked into the clusterInstallerProject/src/main/resources up to date and the
one that you used to test, Pokuri?

I need to leave now.  I will be back on line just after 10pm.