KAMANJA-407 csv deserialize initial
Initial revision for the deserialized side.

Csv is more or less function complete, but untested.

Caveat: We need to decide how (or if) we are going to support
sparse messages.  Shouldn't it be acceptable to serialize a
subset of the possible fields from a "mapped" message or container
with Csv?

The Json suffers similarly.  Let's discuss later tonight, Pokuri, after
I return.  I will be away for a few hours starting now.