KAMANJA-407 CSV semantics + Corrections
Csv will only support *fixed messages* due to avoid the sullying of the CSV rule that all records have
the same number of columns.  The mapped messages are to be supported by Kamanja's Variable Comma Separated
Value protocol (VCSV).  There is a note about that in the CSV subtask to KAMANJA-407.

The wrong field array was being enumerated on both Json and Csv implementations.  This has been fixed.

The Json serializer deserializer will write only those MappedMsgTypeDef's columns that have valid data from
the ContainerInterface instance... thereby supporting 'sparse' data. That is, it is not an error to not
have a value for every field.  Json is really the best way to represent the mapped sparse data at the moment.
When/if the VCSV briefly described above is implemented, then there are choices.

For Json serializer processing of fixed messages, the expectation is that all fields are present. Should one
not be there, an exception is thrown.  This could be relaxed in a future release if desired to allow "null"
or other default values when not all data is present.