KAMANJA-411 Remove AdapterMessageBinding fixes
1) The bindings are deleted, not marked inactive
2) Deletion from persistent store is propagated with zk notification
3) Test case for remove updated.

Notes: No attempt is currently made to see if the adapter message binding is currently in use. Referential
integrity checks are in order.  If a message is removed, the binding is not removed with it and probably should.
Similarly if the adapter is removed, the adapter message binding should also be eliminated.

These are not serious problems, being just unusable bindings, but an issue / jira task should be written up and
implemented in a future sprint.