issue1037 Remove MetadataAPI issues
1) With this checkin, these work:
a) Remove adapter message binding
b) Remove model (remove of jtm model with this)
c)

2) And these partially work:

a) message
b) container

Notes:

The message and container removals are failing because the objects
are not being serialized.  The top of the stack at failure:

java.lang.Exception: serializeObjectToJson doesn't support the objects of type com.ligadata.kamanja.metadata.StructTypeDef yet.
        at com.ligadata.MetadataAPI.MetadataAPISerialization$.serializeObjectToJson(MetadataAPISerialization.scala:633) ~[KamanjaInternalDeps_2.11-1.4.0.jar:0.1-SNAPSHOT]
        at com.ligadata.MetadataAPI.MetadataAPIImpl$.UpdateObjectInDB(MetadataAPIImpl.scala:1037) [KamanjaInternalDeps_2.11-1.4.0.jar:0.1-SNAPSHOT]
        at com.ligadata.MetadataAPI.MetadataAPIImpl$.ModifyObject(MetadataAPIImpl.scala:1487) [KamanjaInternalDeps_2.11-1.4.0.jar:0.1-SNAPSHOT]
        at com.ligadata.MetadataAPI.MetadataAPIImpl$.DeleteObject(MetadataAPIImpl.scala:1516) [KamanjaInternalDeps_2.11-1.4.0.jar:0.1-SNAPSHOT]
        at com.ligadata.MetadataAPI.TypeUtils$.RemoveType(TypeUtils.scala:191) [KamanjaInternalDeps_2.11-1.4.0.jar:0.1-SNAPSHOT]
        at com.ligadata.MetadataAPI.

The is looking for StructTypeDef but we have the object that owns the struct type def ... namely MessageDef in the serializeObjToJson case...

It is sufficient to send the typedef for the messagedef and the containerdef i think.  This code needs to be revised eliminate the wrapper object serialization and jus
t use the type serialization that is there and may be ok.