issue #148 fix md api ingestion
While working on this bug reported by @ligadata-yasser it was discovered that the
output message tag and value were inadverently removed from StartMetadataAPI ... these
are needed for the kpmml and pmml.

To test I added back the PMML test cases that got removed in an earlier edition of the
1.4.0.Test.

Here is sample commands:

$KAMANJA_HOME/bin/kamanja $apiConfigProperties get all messages

$KAMANJA_HOME/bin/kamanja $apiConfigProperties remove message system.irismsg.000000000001000001 TENANTID tenant1

$KAMANJA_HOME/bin/kamanja $apiConfigProperties add message $MetadataDir/message/IrisMsg.json TENANTID tenant1
$KAMANJA_HOME/bin/kamanja $apiConfigProperties add message $MetadataDir/message/IrisMsg1.json TENANTID tenant1

$KAMANJA_HOME/bin/kamanja $apiConfigProperties add model pmml  MODELNAME com.botanical.jpmml.IrisDecisionTree MODELVERSION 000000.000001.000001 MESSAGENAME System.IrisMsg OUTMESSAGE System.IrisMsg1 TENANTID tenant1 $MetadataDir/model/KNIME/DecisionTreeEnsembleIris.pmml

See RemoveUnitTests.md test cases (near end for these tests).