KAMANJA-435 introduce v0 of python model interfaces
... on python side

Notes / General thoughts about the python model:

1) There is an abstract ModelBase class that declares the standard
protocols to be implemented not unlike the scala version.
2) Like the scala version, there is a ModelInstance that implements
the standard methods one would ask of a model (its name, version, et al).
3) Since the Python model is not part of the Kamanja JVM, a proxy
of the information available in the model factory used in the JVM is
supplied to the python server at addModel time.
4) This information includes the standard model information that
the model factory typically give on behalf of the mode instance PLUS
  a) the input messages for the model
  b) output messages for the model
  c) and any other pieces of information that might be needed
5) The message information could be of some compact form (names and types are enough)
6) One might think of this as a temporary measure until the python server has its own
interface to the persistent metadata content for the cluster.
7) The message format sent on the wire for addModel might look like this:

   s"$cmd\n$modelName\n$modelInfo\n$modelSrc"

or perhaps

   s"$cmd\n$modelName\n$modelInfo\n$cntOfInput\n$inputMsgs\n$cntOfOutput\n$outMsgs$modelSrc"

In the first case the input and output messages for the model are part of the modelInfo
dictionary content.  In the 2nd case, these are specified outside (at least for transport to
python server) and then represented in the modelInfo dictionary or separate dictionary(ies)
for that purpose.

Status:

I am changing the SockClient (the test client) and the addModel command on the python side
to accept the 2nd idea (.. this one first anyway).