KAMANJA-435 add ability to send multiple msgs
To test the python server a file may be supplied now to the executeModel command:

	SocketClient.scala --modelName '<modelName>' --filePath '<msg file path>'

All messages are sent at once in this version.  A json string is returned for all results.
The executeModel is where the server side changes occurred.

A test file has been added to test the arithmetic "models".

A bit more documentation in the README.md