KAMANJA-435 Fixed divide.py
Models add.py, multiply.py, divide.py tested with addModel.

1) Apparently the console output I was looking at gets disconnected on the server window.
The models can be added one after the other with the script but one does not see the
server side console logging.

2) The fix for this is to introduce the log4j package in the server.  As we discussed,
it looks like we can ultimately pass the logging config file to the pythonserver.  What
might be an issue is that the level of the software being used in our scala components
might be too high a level for the 2.7x python.

3. small fix to the sock client.

4. the python server is, btw, startable with the sock client (See SockClientsTest.md for
some examples).  It will print the pid to the console of the python server started (a
ps aux | grep | sed script is invoked with a Process for this purpose).